# Conflicts:
#	data/03.会社の初期値コピー/0会社初期値マスタCSV/SSPCT_MASTERCOPY_DATA.csv

+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.app/src/main/java/nts/uk/ctx/at/record/app/find/monthly/root/dto/PremiumTimeOfMonthlyDto.java
@@ -24,7 +24,7 @@ public class PremiumTimeOfMonthlyDto implements ItemConst, AttendanceItemDataGat
 	private List<AggregatePremiumTimeDto> premiumTimes;
 	
 	/**　割増金額合計*/
-	private int premiumAmountTotal;
+	private long premiumAmountTotal;
 
 	public static PremiumTimeOfMonthlyDto from(PremiumTimeOfMonthly domain) {
 		PremiumTimeOfMonthlyDto dto = new PremiumTimeOfMonthlyDto();
